+++ ElDorito/Source/Blam/BlamPlayers.hpp
@@ -30,8 +30,11 @@ namespace Blam::Players
 		{
 			Helmet = 0,
 			Chest,
-			RightShoulder,
-			LeftShoulder,
+			Shoulders,
+			Arms,
+			Legs,
+			Acc,
+			Pelvis,
 
 			Count
 		};
@@ -42,9 +45,6 @@ namespace Blam::Players
 	{
 		uint32_t Colors[ColorIndices::Count];
 		uint8_t Armor[ArmorIndices::Count];
-		uint8_t : 8;
-		uint8_t : 8;
-		uint8_t : 8;
 		uint32_t Unknown1C;
 	};
 	static_assert(sizeof(PlayerCustomization) == 0x20, "Invalid PlayerCustomization size");
