+++ ElDorito/Source/Modules/ModulePlayer.cpp
@@ -1,7 +1,7 @@
 #include <ctime>
 #include "ModulePlayer.hpp"
 #include "../ElDorito.hpp"
-#include "../Game/Armor.hpp"
+#include "../Patches/Armor.hpp"
 #include "../Patches/PlayerRepresentation.hpp"
 #include "../Patches/PlayerUid.hpp"
 #include <iomanip>
@@ -10,7 +10,7 @@ namespace
 {
 	bool VariablePlayerArmorUpdate(const std::vector<std::string>& Arguments, std::string& returnInfo)
 	{
-		Game::Armor::RefreshUiPlayer();
+		Patches::Armor::RefreshUiPlayer();
 		return true;
 	}
 
@@ -29,7 +29,7 @@ namespace
 #ifdef _DEBUG
 		Patches::PlayerRepresentation::UpdateLocalRepresentation();
 #endif
-		Game::Armor::RefreshUiPlayer();
+		Patches::Armor::RefreshUiPlayer();
 		return true;
 	}
 
@@ -62,7 +62,7 @@ namespace
 			return false;
 		}
 
-		Game::Armor::SetUiPlayerModelTransform(&position, nullptr);
+		Patches::Armor::SetUiPlayerModelTransform(&position, nullptr);
 
 		return true;
 	}
@@ -76,7 +76,7 @@ namespace
 			return false;
 		}
 
-		Game::Armor::SetUiPlayerModelTransform(nullptr, &rotationAngle);
+		Patches::Armor::SetUiPlayerModelTransform(nullptr, &rotationAngle);
 
 		return true;
 	}
@@ -86,10 +86,13 @@ namespace Modules
 {
 	ModulePlayer::ModulePlayer() : ModuleBase("Player")
 	{
-		VarArmorHelmet = AddVariableString("Armor.Helmet", "armor_helmet", "Armor ID for player helmet", eCommandFlagsArchived, "", VariablePlayerArmorUpdate);
+		VarArmorAccessory = AddVariableString("Armor.Accessory", "armor_accessory", "Armor ID for player accessory", eCommandFlagsArchived, "", VariablePlayerArmorUpdate);
+		VarArmorArms = AddVariableString("Armor.Arms", "armor_arms", "Armor ID for player arms", eCommandFlagsArchived, "", VariablePlayerArmorUpdate);
 		VarArmorChest = AddVariableString("Armor.Chest", "armor_chest", "Armor ID for player chest", eCommandFlagsArchived, "", VariablePlayerArmorUpdate);
-		VarArmorRightShoulder = AddVariableString("Armor.RightShoulder", "armor_right_shoulder", "Armor ID for player right shoulder", eCommandFlagsArchived, "", VariablePlayerArmorUpdate);
-		VarArmorLeftShoulder = AddVariableString("Armor.LeftShoulder", "armor_left_shoulder", "Armor ID for player left shoulder", eCommandFlagsArchived, "", VariablePlayerArmorUpdate);
+		VarArmorHelmet = AddVariableString("Armor.Helmet", "armor_helmet", "Armor ID for player helmet", eCommandFlagsArchived, "", VariablePlayerArmorUpdate);
+		VarArmorLegs = AddVariableString("Armor.Legs", "armor_legs", "Armor ID for player legs", eCommandFlagsArchived, "", VariablePlayerArmorUpdate);
+		VarArmorPelvis = AddVariableString("Armor.Pelvis", "armor_pelvis", "Armor ID for player pelvis", eCommandFlagsArchived, "", VariablePlayerArmorUpdate);
+		VarArmorShoulders = AddVariableString("Armor.Shoulders", "armor_shoulders", "Armor ID for player shoulders", eCommandFlagsArchived, "", VariablePlayerArmorUpdate);
 		AddCommand("Armor.Update", "armor_update", "Update the player's armor.", eCommandFlagsHidden, VariablePlayerArmorUpdate);
 		AddCommand("Armor.SetUiModelPosition", "armor_ui_player_model_position", "Set the position of the ui player model", (CommandFlags)(eCommandFlagsOmitValueInList | eCommandFlagsHidden), CommandSetUiPlayerModelPosition);
 		AddCommand("Armor.SetUiModelRotation", "armor_ui_player_model_rotation", "Set the rotation of the ui player model", (CommandFlags)(eCommandFlagsOmitValueInList | eCommandFlagsHidden), CommandSetUiPlayerModelRotation);
