+++ ElDorito/Source/Modules/ModulePlayer.hpp
@@ -7,10 +7,13 @@ namespace Modules
 	class ModulePlayer : public Utils::Singleton<ModulePlayer>, public ModuleBase
 	{
 	public:
-		Command* VarArmorHelmet;
+		Command* VarArmorAccessory;
+		Command* VarArmorArms;
 		Command* VarArmorChest;
-		Command* VarArmorRightShoulder;
-		Command* VarArmorLeftShoulder;
+		Command* VarArmorHelmet;
+		Command* VarArmorLegs;
+		Command* VarArmorPelvis;
+		Command* VarArmorShoulders;
 
 		Command* VarColorsPrimary;
 		Command* VarColorsSecondary;
